requirements.txt
Django==2.2.16
mixer==7.1.2
Pillow==8.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
django-debug-toolbar==3.2.4
pytils
flake8
mypy
flake8-broken-line
flake8-return
flake8-isort
pep8-naming
mypy
coverage
djangorestframework
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
djoser==2.1.0
django_allauth==0.51.0
python-dotenv==0.19.0
